iOS: handle images pasted properly
